Updated the build scripts to allow for reruns

+++ scripts/setup_ws.sh
@@ -35,12 +35,26 @@ sudo -H pip install -U \
 # Create the workspace directories
 mkdir -p ./stable/src ./active/src
 
-# Clone the repo
-git clone -b dev git@github.com:gt-rail-internal/derail-fetchit.git ./active/src/derail-fetchit
+# If the repo has not been cloned, then clone it
+if [ ! -d ./active/src/derail-fetchit ]; then
+    git clone -b dev git@github.com:gt-rail-internal/derail-fetchit.git ./active/src/derail-fetchit
+else
+    echo "$(tput bold)Repository exists; Not cloning$(tput sgr0)"
+fi
 
-# Setup the symlinks and initialize the workspaces
-ln -s $(pwd)/active/src/derail-fetchit/scripts/rosinstall/active.rosinstall ./active/src/.rosinstall
-ln -s $(pwd)/active/src/derail-fetchit/scripts/rosinstall/stable.rosinstall ./stable/src/.rosinstall
+# If the rosinstall files don't exist, setup the symlinks
+if [ ! -f ./active/src/.rosinstall ]; then
+    ln -s $(pwd)/active/src/derail-fetchit/scripts/rosinstall/active.rosinstall ./active/src/.rosinstall
+else
+    echo "$(tput bold)Active workspace rosinstall file exists; Not linking$(tput sgr0)"
+fi
+if [ ! -f ./stable/src/.rosinstall ]; then
+    ln -s $(pwd)/active/src/derail-fetchit/scripts/rosinstall/stable.rosinstall ./stable/src/.rosinstall
+else
+    echo "$(tput bold)Stable workspace rosinstall file exists; Not linking$(tput sgr0)"
+fi
+
+# Initialize the workspaces
 cd ./stable/src/ && wstool up && cd ../../
 cd ./active/src/ && wstool up && cd ../../
